refactored scraper and created module
