+++ config/application.rb
@@ -10,6 +10,7 @@ module BProjectEther
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.1
+    config.autoload_paths += %W(#{config.root}/lib) # add this line
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
