+++ config/routes.rb
@@ -1,10 +1,11 @@
 Rails.application.routes.draw do
   mount_devise_token_auth_for 'User', at: 'api/auth'
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
-  get '/zmartPS4', to: 'scraping#zmart_ps4_scrape'
-  get '/zmartXBONE', to: 'scraping#zmart_xbone_scrape'
-  get '/weplayPS4', to: 'scraping#weplay_ps4_scrape'
-  get '/weplayXBONE', to: 'scraping#weplay_xbone_scrape'
-  get '/sniperPS4', to: 'scraping#sniper_ps4_scrape'
-  get '/sniperXBONE', to: 'scraping#sniper_xbone_scrape'
+  get '/zmartPS4', to: 'scraping#zmart_ps4'
+  get '/zmartXBONE', to: 'scraping#zmart_xbone'
+  get '/weplayPS4', to: 'scraping#weplay_ps4'
+  get '/weplayXBONE', to: 'scraping#weplay_xbone'
+  get '/sniperPS4', to: 'scraping#sniper_ps4'
+  get '/sniperXBONE', to: 'scraping#sniper_xbone'
 end
+
